v3.8.4: bitwise shift fixes, array concat, range step, io.readbytes non-seekable
Core bugfixes across all backends (interpreter, C codegen, AOT, LLVM, VM):
- >> and << now use unsigned arithmetic (fixes sign-extension on values >= 2^63)
- [1,2] + [3,4] now returns [1,2,3,4] instead of nil
- range(start, end, step) accepts 3 arguments with positive/negative step
- io.readbytes reads non-seekable files (e.g. /dev/urandom) via chunked fallback

+++ core/src/c/llvm_backend.c
@@ -591,8 +591,8 @@ static ImportConstValue llvm_eval_const_expr(Expr* expr, ModuleConst* consts, in
                     if (op.type == TOKEN_AMP) result = import_const_number((double)(li & ri));
                     if (op.type == TOKEN_PIPE) result = import_const_number((double)(li | ri));
                     if (op.type == TOKEN_CARET) result = import_const_number((double)(li ^ ri));
-                    if (op.type == TOKEN_LSHIFT) result = import_const_number((double)(li << ri));
-                    if (op.type == TOKEN_RSHIFT) result = import_const_number((double)(li >> ri));
+                    if (op.type == TOKEN_LSHIFT) result = import_const_number((double)((unsigned long long)li << ri));
+                    if (op.type == TOKEN_RSHIFT) result = import_const_number((double)((unsigned long long)li >> ri));
                 }
             }
